fixed graph statistics
